# contracts/openapi/classduck-api.v1.yaml
openapi: 3.0.3
info:
  title: ClassDuck API
  version: 1.1.0
  description: |
    上课鸭前后端接口契约。
    - /v1 路径为 Flutter 兼容接口（对接 Python 或 Node 后端均可）。
    - /api 路径为 Python 导入后端新增接口。
servers:
  - url: http://localhost:8000
    description: Python import service (dev)
  - url: https://api.classduck.local
    description: Production
paths:
  /v1/config/schools:
    get:
      summary: Get school import configurations
      operationId: getSchoolConfigs
      responses:
        '200':
          description: school config list
          content:
            application/json:
              schema:
                $ref: ../schemas/school-config-list.schema.json
        '500':
          description: server error
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.schema.json
  /v1/config/adapters:
    get:
      summary: Get parser adapter rules
      operationId: getAdapterRules
      responses:
        '200':
          description: adapter rule list
          content:
            application/json:
              schema:
                $ref: ../schemas/adapter-rule-list.schema.json
        '500':
          description: server error
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.schema.json
  /v1/import/logs:
    post:
      summary: Upload import failure log
      operationId: postImportLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/import-log-request.schema.json
      responses:
        '202':
          description: accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  traceId:
                    type: string
                required:
                  - accepted
                  - traceId
        '400':
          description: invalid payload
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.schema.json
  /v1/release/check:
    get:
      summary: Check latest app release
      operationId: checkRelease
      parameters:
        - name: currentVersion
          in: query
          required: true
          schema:
            type: string
        - name: platform
          in: query
          required: true
          schema:
            type: string
            enum: [android, ios]
      responses:
        '200':
          description: release check result
          content:
            application/json:
              schema:
                $ref: ../schemas/release-check-response.schema.json
        '500':
          description: server error
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.schema.json

  /api/schools:
    get:
      summary: List all schools (new API)
      operationId: listSchools
      tags: [import-service]
      responses:
        '200':
          description: school summary list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    system_type:
                      type: string

  /api/schools/{schoolId}/config:
    get:
      summary: Get full school configuration
      operationId: getSchoolConfig
      tags: [import-service]
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: full school config (field_mapping, timer_config, etc.)
          content:
            application/json:
              schema:
                type: object
        '404':
          description: school not found

  /api/schools/{schoolId}/script:
    get:
      summary: Download JS provider script for a school
      operationId: getSchoolScript
      tags: [import-service]
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
        - name: script_type
          in: query
          schema:
            type: string
            default: provider
      responses:
        '200':
          description: JS script content
          content:
            application/json:
              schema:
                type: object
                properties:
                  school_id:
                    type: string
                  script_type:
                    type: string
                  script:
                    type: string
        '404':
          description: script not found

  /api/import/validate:
    post:
      summary: Validate and parse raw schedule JSON
      operationId: validateImport
      tags: [import-service]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [school_id, raw_data]
              properties:
                school_id:
                  type: string
                raw_data:
                  type: string
                  description: JSON string of raw course data array
                year:
                  type: string
                term:
                  type: string
      responses:
        '200':
          description: standardized import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      semester_name:
                        type: string
                      school_id:
                        type: string
                      courses:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            position:
                              type: string
                            teacher:
                              type: string
                            weeks:
                              type: array
                              items:
                                type: integer
                            day:
                              type: integer
                            start_section:
                              type: integer
                            duration:
                              type: integer
                      total_count:
                        type: integer
                      valid_count:
                        type: integer
                      invalid_count:
                        type: integer
                      warnings:
                        type: array
                        items:
                          type: string

  /api/import/log:
    post:
      summary: Report import log (new API)
      operationId: reportImportLog
      tags: [import-service]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [school_id, status]
              properties:
                school_id:
                  type: string
                status:
                  type: string
                  enum: [success, fail]
                error_message:
                  type: string
                course_count:
                  type: integer
      responses:
        '200':
          description: log accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean

  /health:
    get:
      summary: Health check
      operationId: healthCheck
      tags: [system]
      responses:
        '200':
          description: service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

components: {}